search customers except history
