+++ restapi/controllers/controllers.py
@@ -61,6 +61,7 @@ class Restapi(http.Controller):
             'image':base64.b64decode(image) if image else '',
             }
     
+    
     def get_params(self,url):
         params = url.split('?')
         params = params[1].split('&')
